Move deadlock tests to separate file
This lets us disable the deadlock tests in the race detector, keeping CI
clean. Unfortunately, the deadlock library in use does not allow atomic
control over these values and doesn't recommend setting them
post-startup. While running in tests are serial and nothing else is
executing in the process state, it is not ideal but workable. This code
is not used in production.
